[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.68 ms, Memory: 81.1 MB -BaekjoonHub

+++ 프로그래머스/lv1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -9,13 +9,13 @@ class Solution {
         for(int i=0;i<n;i++){
             list.add(num);
      
-            num += (long)x;  
+            num += x;  
         }
         
         long[] arr = new long[list.size()];
         
         for(int i=0;i<arr.length;i++){
-            arr[i] = (long)list.get(i);
+            arr[i] = list.get(i);
         }
